+++ prisma/schema.prisma
@@ -3,27 +3,42 @@ generator client {
 }
 
 datasource db {
-  provider = "postgresql"
+  provider  = "postgresql"
   url       = env("POSTGRES_PRISMA_URL") // uses connection pooling
   directUrl = env("POSTGRES_URL_NON_POOLING") // uses a direct connection
 }
 
 model users {
-  id          Int    @id @default(autoincrement())
-  email       String @unique(map: "unique_email")
-  password    String?
-  savedBills  savedBill[]
+  id            Int            @id @default(autoincrement())
+  email         String         @unique(map: "unique_email")
+  password      String?
+  savedBills    savedBill[]
+  savedSponsors savedSponsor[]
 }
 
 model savedBill {
-  id                  Int     @id @default(autoincrement())
-  userId              Int
-  legiscanBillId      Int
-  title               String
-  description         String
-  state               String
-  sessionTitle        String
-  sessionId           Int
-  changeHash          String
-  user                users   @relation(fields: [userId], references: [id])
+  id             Int    @id @default(autoincrement())
+  userId         Int
+  legiscanBillId Int
+  title          String
+  description    String
+  state          String
+  sessionTitle   String
+  sessionId      Int
+  changeHash     String
+  user           users  @relation(fields: [userId], references: [id])
+}
+
+model savedSponsor {
+  id               Int     @id @default(autoincrement())
+  userId           Int
+  legiscanPeopleId Int
+  personHash       String
+  stateId          String
+  party            String
+  role             String
+  name             String
+  suffix           String?
+  district         String
+  user             users   @relation(fields: [userId], references: [id])
 }
